requirements.txt
PyYAML
flask
elsa
frozen-flask
markdown
textwrap
jinja2